play: Refactor and clean-up.
